+++ salt/utils/__init__.py
@@ -849,10 +849,13 @@ def required_module_list(docstring=None):
     Return a list of python modules required by a salt module that aren't
     in stdlib and don't exist on the current pythonpath.
     '''
+    # Late import to avoid circular import.
+    import salt.utils.doc
+
     if not docstring:
         return []
     ret = []
-    modules = parse_docstring(docstring).get('deps', [])
+    modules = salt.utils.doc.parse_docstring(docstring).get('deps', [])
     for mod in modules:
         try:
             if six.PY3:
@@ -2191,30 +2194,18 @@ def parse_docstring(docstring):
             'full': full docstring,
             'deps': list of dependencies (empty list if none)
         }
+
+    .. deprecated:: Oxygen
     '''
-    # First try with regex search for :depends:
-    ret = {}
-    ret['full'] = docstring
-    regex = r'([ \t]*):depends:[ \t]+- (\w+)[^\n]*\n(\1[ \t]+- (\w+)[^\n]*\n)*'
-    match = re.search(regex, docstring, re.M)
-    if match:
-        deps = []
-        regex = r'- (\w+)'
-        for line in match.group(0).strip().splitlines():
-            deps.append(re.search(regex, line).group(1))
-        ret['deps'] = deps
-        return ret
-    # Try searching for a one-liner instead
-    else:
-        txt = 'Required python modules: '
-        data = docstring.splitlines()
-        dep_list = list(x for x in data if x.strip().startswith(txt))
-        if not dep_list:
-            ret['deps'] = []
-            return ret
-        deps = dep_list[0].replace(txt, '').strip().split(', ')
-        ret['deps'] = deps
-        return ret
+    warn_until(
+        'Neon',
+        'Use of \'salt.utils.parse_docstring\' detected. This function has been moved to '
+        '\'salt.utils.doc.parse_docstring\' as of Salt Oxygen. This warning will be '
+        'removed in Salt Neon.'
+    )
+    # Late import to avoid circular import.
+    import salt.utils.doc
+    return salt.utils.doc.parse_docstring(docstring)
 
 
 def print_cli(msg, retries=10, step=0.01):
